+++ apps/web/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @stoker-platform/web-app
 
+## 0.5.106
+
+### Patch Changes
+
+- @stoker-platform/node-client@0.5.59
+- @stoker-platform/utils@0.5.50
+- @stoker-platform/web-client@0.5.59
+
 ## 0.5.105
 
 ### Patch Changes