+++ packages/cli/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @stoker-platform/cli
 
+## 0.5.83
+
+### Patch Changes
+
+- feat: add collection TypeScript types generation
+- Updated dependencies
+    - @stoker-platform/types@0.5.38
+    - @stoker-platform/node-client@0.5.59
+    - @stoker-platform/utils@0.5.50
+
 ## 0.5.82
 
 ### Patch Changes